Renaming the shame subcommand for namespace issue
