+++ DiscordBot/Commands/Interactive2/Graveyard/ShameSubCommand/ShameAutoCompleteRequest.cs
@@ -1,7 +1,7 @@
 using DiscordBot.Commands.Interactive2.Base.Requests;
 using DiscordBot.Common.Models.Enums;
 
-namespace DiscordBot.Commands.Interactive2.Graveyard.Shame;
+namespace DiscordBot.Commands.Interactive2.Graveyard.ShameSubCommand;
 
 public class ShameAutoCompleteRequest :AutoCompleteCommandRequestBase<ShameSubCommandDefinition>{
 	public ShameAutoCompleteRequest(AutocompleteCommandContext context) : base(context) { }
