+++ DiscordBot/Commands/Interactive2/Graveyard/ShameSubCommand/ShameCommandRequest.cs
@@ -1,7 +1,7 @@
 using DiscordBot.Commands.Interactive2.Base.Requests;
 using DiscordBot.Common.Models.Enums;
 
-namespace DiscordBot.Commands.Interactive2.Graveyard.Shame; 
+namespace DiscordBot.Commands.Interactive2.Graveyard.ShameSubCommand; 
 
 public class ShameCommandRequest :ApplicationCommandRequestBase<ShameSubCommandDefinition>{
 	public ShameCommandRequest(ApplicationCommandContext context) : base(context) { }
